+++ alertaclient/commands/cmd_notes.py
@@ -5,14 +5,14 @@ from tabulate import tabulate
 
 
 @click.command('notes', short_help='List notes')
-@click.option('--alert-id', '-i', metavar='UUID', help='alert IDs (can use short 8-char id)')
+@click.option('--alert-id', '-i', metavar='ID', help='alert IDs (can use short 8-char id)')
 @click.pass_obj
 def cli(obj, alert_id):
     """List notes."""
     client = obj['client']
     if alert_id:
         if obj['output'] == 'json':
-            r = client.http.get('/alert/{}/notes'.format(alert_id))
+            r = client.http.get(f'/alert/{alert_id}/notes')
             click.echo(json.dumps(r['notes'], sort_keys=True, indent=4, ensure_ascii=False))
         else:
             timezone = obj['timezone']