+++ alertaclient/models/heartbeat.py
@@ -1,3 +1,4 @@
+import re
 from datetime import datetime, timedelta
 
 from alertaclient.utils import DateTime
@@ -46,6 +47,14 @@ class Heartbeat:
         if not isinstance(json.get('timeout', 0), int):
             raise ValueError('timeout must be an integer')
 
+        create_time = json.get('createTime')
+        if create_time is not None:
+            create_time = re.sub(r'(\d{3})(\d*)(Z)', r'\1\3', create_time)
+
+        receiveTime = json.get('receiveTime')
+        if receiveTime is not None:
+            receiveTime = re.sub(r'(\d{3})(\d*)(Z)', r'\1\3', receiveTime)
+
         return Heartbeat(
             id=json.get('id', None),
             origin=json.get('origin', None),
@@ -53,10 +62,10 @@ class Heartbeat:
             tags=json.get('tags', list()),
             attributes=json.get('attributes', dict()),
             event_type=json.get('type', None),
-            create_time=DateTime.parse(json.get('createTime')),
+            create_time=DateTime.parse(create_time),
             timeout=json.get('timeout', None),
             max_latency=json.get('maxLatency', None) or DEFAULT_MAX_LATENCY,
-            receive_time=DateTime.parse(json.get('receiveTime')),
+            receive_time=DateTime.parse(receiveTime),
             customer=json.get('customer', None)
         )
 
@@ -70,8 +79,8 @@ class Heartbeat:
             'createTime': DateTime.localtime(self.create_time, timezone),
             'receiveTime': DateTime.localtime(self.receive_time, timezone),
             'since': self.since,
-            'timeout': '{}s'.format(self.timeout),
-            'latency': '{:.0f}ms'.format(self.latency),
-            'maxLatency': '{}ms'.format(self.max_latency),
+            'timeout': f'{self.timeout}s',
+            'latency': f'{self.latency:.0f}ms',
+            'maxLatency': f'{self.max_latency}ms',
             'status': self.status
         }